% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{get_card_by_id}
\alias{get_card_by_id}
\title{Retrieve a card from the Scryfall API using the card's ID.}
\usage{
get_card_by_id(
  id,
  type = "scryfall",
  format = NULL,
  face = NULL,
  version = NULL,
  set = NULL
)
}
\arguments{
\item{id}{ID value(s) to search for.}

\item{type}{Which ID type to use ("scryfall" (default), "mtgo", "arena", "collector", "multiverse").}

\item{format}{Which format should be retrieved.}

\item{face}{Which card face to return.}

\item{version}{Which image version to return ("" (default), "small", "normal", "large").}

\item{set}{Which set to request (Optional).}
}
\value{
A tibble containing card information.
}
\description{
Retrieve a card from the Scryfall API using the card's ID.
}
