% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{get_card_by_name}
\alias{get_card_by_name}
\title{Retrieve a card from the Scryfall API using the card's name.}
\usage{
get_card_by_name(name, fuzzy = FALSE, set = NULL)
}
\arguments{
\item{name}{The card name to be searched for.}

\item{fuzzy}{Whether to search for an exact match, or to use a fuzzy search.}

\item{set}{(Optional) A set code to limit the search to one set.}
}
\value{
A tibble containing card information.
}
\description{
For both exact and fuzzy, card names are case-insensitive and punctuation is
optional (you can drop apostrophes and periods etc). For example: fIReBALL is
the same as Fireball and smugglers copter is the same as Smuggler's Copter.
}
