% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_set_by_code}
\alias{get_set_by_code}
\title{Get an individual set based on its 3-letter identifier}
\usage{
get_set_by_code(code)
}
\arguments{
\item{code}{The set's 3-letter code}
}
\value{
A tibble containing the requested set
}
\description{
Get an individual set based on its 3-letter identifier
}
