% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{get_set_by_id}
\alias{get_set_by_id}
\title{Retrieve a set by an ID value.}
\usage{
get_set_by_id(id, type = c("scryfall"))
}
\arguments{
\item{id}{the set's ID.}

\item{type}{the type of ID to use in the search.}
}
\description{
Retrieve a set by an ID value.
}
