% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{search_cards}
\alias{search_cards}
\title{Search for cards based on scryfall's search syntax.}
\usage{
search_cards(
  q,
  unique = "cards",
  order = "name",
  dir = "auto",
  include_extras = FALSE,
  page = 1
)
}
\arguments{
\item{q}{Query, using scryfall's syntax.}

\item{unique}{Whether scryfall should remove possible duplicates in the results ("cards" (default), "art", "prints")}

\item{order}{Method for how the returned cards should be sorted.}

\item{dir}{The direction to sort cards.}

\item{include_extras}{When TRUE, extra cards (e.g. tokens) will be included (default: FALSE).}

\item{page}{Which page number to return; defaults to 1.}
}
\value{
A tibble containing card information.

Scryfall's search syntax reference can be \href{https://scryfall.com/docs/syntax}{found here.} Below is further information on different options that can be specified for the function arguments.

\emph{\strong{\code{unique}}}

Further information on \code{unique}, from scryfall's documentation:cards (default) Removes duplicate gameplay objects (cards that share a name and have the same functionality). For example, if your search matches more than one print of Pacifism, only one copy of Pacifism will be returned.art Returns only one copy of each unique artwork for matching cards. For example, if your search matches more than one print of Pacifism, one card with each different illustration for Pacifism will be returned, but any cards that duplicate artwork already in the results will be omitted.prints Returns all prints for all cards matched (disables rollup). For example, if your search matches more than one print of Pacifism, all matching prints will be returned.

\emph{\strong{\code{order}}}

Further information on \code{order}, from scryfall's documentation:
\itemize{
\item \emph{name} (default) Sort cards by name, A → Z
\item \emph{set} Sort cards by their set and collector number: AAA/#1 → ZZZ/#999
\item \emph{released} Sort cards by their release date: Newest → Oldest
\item \emph{rarity} Sort cards by their rarity: Common → Mythic
\item \emph{color} Sort cards by their color and color identity: WUBRG → multicolor → colorless
\item \emph{usd} Sort cards by their lowest known U.S. Dollar price: 0.01 → highest, null last
\item \emph{tix} Sort cards by their lowest known TIX price: 0.01 → highest, null last
\item \emph{eur} Sort cards by their lowest known Euro price: 0.01 → highest, null last
\item \emph{cmc} Sort cards by their converted mana cost: 0 → highest
\item \emph{power} Sort cards by their power: null → highest
\item \emph{toughness} Sort cards by their toughness: null → highest
\item \emph{edhrec} Sort cards by their EDHREC ranking: lowest → highest
\item \emph{artist} Sort cards by their front-side artist name: A → Z
}

\emph{\strong{\code{dir}}}
\itemize{
\item \emph{auto} (default) Scryfall will automatically choose the most inuitive direction to sort
\item \emph{asc} Sort ascending (the direction of the arrows in the previous table)
\item \emph{desc} Sort descending (flip the direction of the arrows in the previous table)
}
}
\description{
Scryfall's API returns a maximum of 175 cards per page.
}
