% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{unpack_card_response}
\alias{unpack_card_response}
\title{Return card API results as a data frame.}
\usage{
unpack_card_response(card_content)
}
\arguments{
\item{card_content}{List of card information to be converted into a tibble.}
}
\value{
A tibble with unpacked card information.
}
\description{
Return card API results as a data frame.
}
